requirements.txt
asgiref==3.2.10
Django==3.1.1
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==2.3.0
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
